new controllor to get api endpoints update on postman collection update on questionnareAnswer services

+++ Code/services/src/main/kotlin/pt/iselearning/services/domain/questionnaires/Questionnaire.kt
@@ -23,7 +23,7 @@ data class Questionnaire(
         var description: String?,
 
         @Column(name = "timer", insertable = false, updatable = true)
-        var timer: Int?,
+        var timer: Long?,
 
         @Column(name = "creator_id")
         var creatorId: Int?,
